team_kanban/requirements.txt
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.7.2
django-filter==2.2.0
django-heroku==0.3.1
django-toolbelt==0.0.1
django-widget-tweaks==1.4.5
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.18.1
packaging==20.1
pandas==1.0.1
pep8==1.7.1
Pillow==6.2.1
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pymongo==3.10.1
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.2
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
termcolor==1.1.0
typed-ast==1.4.1
urllib3==1.25.7
wcwidth==0.1.8
whitenoise==3.3.1
wrapt==1.11.2
zipp==3.0.0
